requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=0.19.0
langchain
langchain-openai
langchain-azure-ai
langchain-core
motor==3.3.1
pymongo==4.5.0
python-multipart
httpx>=0.25.0
azure-cognitiveservices-vision-computervision>=0.9.0
msrest>=0.7.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests 